Weaver throttling time should be u64, not u32
This adds a new field to the ReadResponse message in the
weaver.proto file. It keeps the old field too, since we need to
support both old and new values for a while for compatiblity, but
renames the old field for clarity.

However it *changes* the `struct nos2_weaver_read_response`
definition since that's no longer used to communicate between the
AP and GSC and is only used internally by GSC firmware. That API
is deprecated and will be removed from this repo in a future CL.

Bug: 394054536
Bug: 394054538
Test: nugget_targeted_tests

  In addition to running the tests with both GSC firmware and AP
  source changes together, I tested them separately (new GSC +
  old AP, old GSC + new AP) to be sure that returning/expecting
  both u32 and u64 values from protobuf worked correctly.

Change-Id: I7c27851d41014fea2ae87459af6fbc0b08b28af9
Signed-off-by: Bill Richardson <wfrichar@google.com>
Reviewed-on: https://nugget-os-review.googlesource.com/c/host/generic/+/74965
Presubmit-Verified: Treehugger Robot <android-test-infra-workplan-finisher@system.gserviceaccount.com>
Reviewed-by: Kevin Zhang <zhakevin@google.com>